Code refinements in Bitmap class + different errors for when a bitmap file was not found vs. when its format is not supported

+++ Oxygen/soncthrickles/source/sonic3air/menu/GameMenuBase.cpp
@@ -349,7 +349,7 @@ bool GameMenuEntries::sanitizeSelectedIndex(bool allowNonInteractableEntries)
 		if (mEntries.empty())
 			return false;
 
-		mSelectedEntryIndex = mEntries.size() - 1;
+		mSelectedEntryIndex = (int)mEntries.size() - 1;
 	}
 
 	if (!allowNonInteractableEntries && !mEntries[mSelectedEntryIndex]->isFullyInteractable())
